Player with less than 100 will no longer pay patent fees. Fixed patent registration cost. Fixed showing company name correctly in wage messages. Fixed default permission for /patent apply

+++ src/com/dogonfire/seriousbusiness/commands/CommandInfo.java
@@ -75,15 +75,15 @@ public class CommandInfo extends SeriousBusinessCommand
 		
 		if(report.stockValueChange > 0)
 		{
-			sender.sendMessage(ChatColor.AQUA + " StockValue: " + report.stockEndValue + ChatColor.GREEN + "+" + df.format(100 * (report.stockValueChange / report.stockStartValue)) + "%");
+			sender.sendMessage(ChatColor.AQUA + " StockValue: " + (int)report.stockEndValue + ChatColor.GREEN + "+" + df.format(100 * (report.stockValueChange / report.stockStartValue)) + "%");
 		}
 		else if(report.stockValueChange == 0)
 		{
-			sender.sendMessage(ChatColor.AQUA + " StockValue: " + report.stockEndValue);			
+			sender.sendMessage(ChatColor.AQUA + " StockValue: " + (int)report.stockEndValue);			
 		}		
 		else
 		{
-			sender.sendMessage(ChatColor.AQUA + " StockValue: " + report.stockEndValue + ChatColor.RED + df.format(100 * (report.stockValueChange / report.stockStartValue)) + "%");			
+			sender.sendMessage(ChatColor.AQUA + " StockValue: " + (int)report.stockEndValue + ChatColor.RED + df.format(100 * (report.stockValueChange / report.stockStartValue)) + "%");			
 		}		
 		
 		sender.sendMessage(ChatColor.YELLOW + "People:");
